Add date to uploaded filename
Renames handleUpload to handleAttachment because it now also attaches the upload to the email. The reason it does this is so that it can use the original filename... which it doesn't necessarily have to do from there but I think it's tidier.
Prepends time and date in format ssmmhhDDMMYY-filename to the uploaded file, so that dupes aren't overwritten. I didn't store the file names by date in a folder, because I expect that dupes could be uploaded on the same day and doing it more granularly could be overkill.
Cool.