Function to add email to database
This will store the message details, with or without the attachment name. I chose to use the attachment name stored in the uploads folder, rather than the attachment name sent in the email, to make auditing etc easier. It returns the id for use when storing (in the linking table) the users/emails and whether they sent successfully or not.